refactor(daengle): 코드리뷰를 토대로 빠진 api 경로 수정 및 폴더구조 수정
