+++ apps/daengle/src/components/mypage/pet-profile/chip-toggle-button/index.styles.tsx
@@ -13,14 +13,14 @@ export const wrapper = ({
 }) => css`
   ${size === 'fixed' &&
   css`
-    width: 67px;
-    height: 38px;
-
     display: flex;
     align-items: center;
     justify-content: center;
 
+    width: 67px;
+    height: 38px;
     border-radius: 19px;
+
     ${theme.typo.body11};
   `}
 
@@ -33,14 +33,14 @@ export const wrapper = ({
 
    ${size === 'full' &&
   css`
-    width: 100%;
-    padding: 10px 0;
-
     display: flex;
     align-items: center;
     justify-content: center;
 
+    width: 100%;
+    padding: 10px 0;
     border-radius: 28px;
+
     ${theme.typo.body10};
   `}
 
@@ -63,8 +63,9 @@ export const wrapper = ({
         color: ${theme.colors.blue200};
       `
     : css`
-        color: ${theme.colors.gray500};
         border: 1px solid ${disabled ? theme.colors.gray300 : theme.colors.gray200};
+
+        color: ${theme.colors.gray500};
 
         /* color: ${theme.colors.black}; */
       `}
